modified user modular
from master brand changed to orm branch, modified to use Flask-SQLAlchemy instead of plain SQL
