+++ flaskr/templates/auth/login.html
@@ -14,8 +14,8 @@
   {% endfor %}
   <form method="post">
     <div>
-      <label for="username">UserName</label>
-      <input type="text" id="username" name="username" value="{{ request.form['username'] }}" autofocus required>
+      <label for="userid">username or email</label>
+      <input type="text" id="userid" name="userid" value="{{ request.form['userid'] }}" autofocus required>
     </div>
     <div>
       <label for="password">Password</label>
